Q1/requirements.txt
altgraph==0.17.2
appdirs==1.4.4
asgiref==3.4.1
astroid==2.7.3
async-generator==1.10
attrs==21.2.0
autopep8==1.6.0
backcall==0.2.0
bleach==4.1.0
certifi==2021.10.8
cffi==1.14.6
charset-normalizer==2.0.9
click==8.0.1
colorama==0.4.4
colorlover==0.3.0
cycler==0.11.0
dataclasses==0.8
decorator==5.0.9
defusedxml==0.7.1
diff-match-patch==20200713
distlib==0.3.2
Django==3.2.6
django-adminplus==0.5
django-bootstrap3==15.0.0
django-crontab==0.7.1
django-import-export==2.5.0
entrypoints==0.3
et-xmlfile==1.1.0
filelock==3.0.12
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
future==0.18.2
greenlet==1.1.1
idna==3.3
importlib-metadata==2.1.1
importlib-resources==5.1.4
ipykernel==5.5.5
ipython==7.16.1
ipython-genutils==0.2.0
isort==5.9.3
itsdangerous==2.0.1
jedi==0.18.0
Jinja2==3.0.1
jsonschema==3.2.0
jupyter-client==7.0.2
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
kiwisolver==1.3.1
lazy-object-proxy==1.6.0
MarkupPy==1.14
MarkupSafe==2.0.1
matplotlib==3.3.4
mccabe==0.6.1
mistune==0.8.4
mysql-connector==2.2.9
nbclient==0.5.4
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
numpy==1.19.5
odfpy==1.4.1
openpyxl==3.0.7
packaging==21.0
pandas==1.1.5
pandocfilters==1.4.3
parso==0.8.2
pefile==2021.9.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
platformdirs==2.3.0
plotly==5.3.1
prompt-toolkit==3.0.20
psutil==5.8.0
psycopg2==2.9.1
psycopg2-binary==2.9.1
ptyprocess==0.7.0
pycodestyle==2.8.0
pycparser==2.20
Pygments==2.10.0
pyinstaller==4.7
pyinstaller-hooks-contrib==2021.4
PyJWT==1.4.2
pylint==2.10.2
pylint-flask==0.6
pylint-flask-sqlalchemy==0.2.0
pylint-plugin-utils==0.7
PyMySQL==1.0.2
pyparsing==2.4.7
pyrsistent==0.18.0
python-dateutil==2.8.2
pytz==2021.1
pywin32==301
pywin32-ctypes==0.2.0
PyYAML==5.4.1
pyzmq==22.2.1
requests==2.26.0
Send2Trash==1.8.0
six==1.16.0
SQLAlchemy==1.4.23
sqlparse==0.4.1
tablib==3.0.0
tenacity==8.0.1
testpath==0.5.0
toml==0.10.2
tornado==6.1
traitlets==4.3.3
typed-ast==1.4.3
typing-extensions==3.10.0.0
urllib3==1.26.7
virtualenv==20.4.7
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.1
wexpect==4.0.0
wrapt==1.12.1
xlrd==2.0.1
XlsxWriter==3.0.2
xlwt==1.3.0
zipp==3.4.1
